# docs/openapi/permesi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "permesi",
    "description": "Identity and Access Management",
    "contact": {
      "name": "Team Permesi",
      "email": "team@permesi.dev"
    },
    "license": {
      "name": "BSD-3-Clause",
      "identifier": "BSD-3-Clause"
    },
    "version": "0.3.43"
  },
  "paths": {
...
    "/v1/auth/session": {
      "get": {
        "tags": [
          "auth"
        ],
        "operationId": "session",
        "responses": {
          "200": {
            "description": "Session is active",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "204": {
            "description": "No active session"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SessionKind": {
        "type": "string",
        "description": "Session kinds used to gate MFA bootstrap and challenge flows.",
        "enum": [
          "full",
          "mfa_bootstrap",
          "mfa_challenge"
        ]
      },
      "SessionResponse": {
        "type": "object",
        "required": [
          "user_id",
          "email",
          "is_operator",
          "session_kind",
          "totp_enabled",
          "webauthn_enabled"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "is_operator": {
            "type": "boolean"
          },
          "session_kind": {
            "$ref": "#/components/schemas/SessionKind"
          },
          "totp_enabled": {
            "type": "boolean"
          },
          "user_id": {
            "type": "string"
          },
          "webauthn_enabled": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
...
}