# services/permesi/src/api/handlers/auth/session.rs
//! Session handlers for active authentication state management.
//!
//! This module provides endpoints for session lookup and logout using the
//! `permesi_session` cookie. It handles the different session kinds and manages
//! cookie generation.

use axum::{
    Json,
    extract::Extension,
    http::{
        HeaderMap, HeaderValue, StatusCode,
        header::{InvalidHeaderValue, SET_COOKIE},
    },
    response::IntoResponse,
};
use sqlx::PgPool;
use std::sync::Arc;
use tracing::error;

use super::{
    admin_storage::operator_enabled,
    session_kind::SessionKind,
    state::AuthState,
    storage::{
        SessionRecord, delete_full_session, delete_mfa_bootstrap_session,
        delete_mfa_challenge_session, lookup_full_session, lookup_mfa_bootstrap_session,
        lookup_mfa_challenge_session,
    },
    types::SessionResponse,
    utils::hash_session_token,
};

const SESSION_COOKIE_NAME: &str = "permesi_session";

#[utoipa::path(
    get,
    path = "/v1/auth/session",
    responses(
        (status = 200, description = "Session is active", body = SessionResponse),
        (status = 204, description = "No active session")
    ),
    tag = "auth"
)]
pub async fn session(headers: HeaderMap, pool: Extension<PgPool>) -> impl IntoResponse {
    // Missing cookies are treated as "no session" to avoid leaking auth state.
    let Some(token) = extract_session_token(&headers) else {
        return StatusCode::NO_CONTENT.into_response();
    };
    match lookup_any_session(&pool, &token).await {
        Ok(Some(SessionRecord {
            user_id,
            email,
            kind,
            ..
        })) => {
            let is_operator = operator_enabled(&pool, user_id).await.unwrap_or(false);

            let mfa_record = super::mfa::storage::load_mfa_state(&pool, user_id)
                .await
                .unwrap_or(None);
            let totp_enabled = mfa_record
                .as_ref()
                .is_some_and(|r| r.recovery_batch_id.is_some());

            let webauthn_enabled = crate::webauthn::SecurityKeyRepo::list_user_keys(&pool, user_id)
                .await
                .is_ok_and(|keys| !keys.is_empty());

            let response = SessionResponse {
                user_id: user_id.to_string(),
                email,
                is_operator,
                session_kind: kind,
                totp_enabled,
                webauthn_enabled,
            };
            (StatusCode::OK, Json(response)).into_response()
        }
        Ok(None) => StatusCode::NO_CONTENT.into_response(),
        Err(err) => {
            error!("Failed to lookup session: {err}");
            StatusCode::INTERNAL_SERVER_ERROR.into_response()
        }
    }
}

/// Resolve a session cookie into a session record, if present.
///
/// Returns `Ok(None)` when the cookie is missing or invalid.
pub(crate) async fn authenticate_session(
    headers: &HeaderMap,
    pool: &PgPool,
) -> Result<Option<SessionRecord>, StatusCode> {
    let Some(token) = extract_session_token(headers) else {
        return Ok(None);
    };
    match lookup_any_session(pool, &token).await {
        Ok(record) => Ok(record),
        Err(err) => {
            error!("Failed to lookup session: {err}");
            Err(StatusCode::INTERNAL_SERVER_ERROR)
        }
    }
}

#[utoipa::path(
    post,
    path = "/v1/auth/logout",
    responses(
        (status = 204, description = "Session cleared")
    ),
    tag = "auth"
)]
pub async fn logout(
    headers: HeaderMap,
    pool: Extension<PgPool>,
    auth_state: Extension<Arc<AuthState>>,
) -> impl IntoResponse {
    if let Some(token) = extract_session_token(&headers)
        && let Err(err) = delete_any_session(&pool, &token).await
    {
        error!("Failed to delete session: {err}");
    }

    // Always clear the cookie, even if the session record was missing.
    let mut response_headers = HeaderMap::new();
    if let Ok(cookie) = clear_session_cookie(auth_state.config()) {
        response_headers.insert(SET_COOKIE, cookie);
    }
    (StatusCode::NO_CONTENT, response_headers).into_response()
}

/// Build a secure `HttpOnly` cookie for the session token with a custom TTL.
pub(crate) fn session_cookie_with_ttl(
    auth_state: &AuthState,
    token: &str,
    ttl_seconds: i64,
) -> Result<HeaderValue, InvalidHeaderValue> {
    // Only mark cookies secure when the frontend is served over HTTPS.
    let secure = auth_state.config().session_cookie_secure();
    let mut cookie = format!(
        "{SESSION_COOKIE_NAME}={token}; Path=/; HttpOnly; SameSite=Lax; Max-Age={ttl_seconds}"
    );
    if secure {
        cookie.push_str("; Secure");
    }
    HeaderValue::from_str(&cookie)
}

/// Build a cookie header that clears the auth session.
///
/// Use this when a flow revokes all sessions so the client drops the cookie.
pub(crate) fn clear_session_cookie(
    auth_config: &super::state::AuthConfig,
) -> Result<HeaderValue, InvalidHeaderValue> {
    let secure = auth_config.session_cookie_secure();
    let mut cookie = format!("{SESSION_COOKIE_NAME}=; Path=/; HttpOnly; SameSite=Lax; Max-Age=0");
    if secure {
        cookie.push_str("; Secure");
    }
    HeaderValue::from_str(&cookie)
}

/// Extract a session token from the `permesi_session` cookie.
pub(crate) fn extract_session_token(headers: &HeaderMap) -> Option<String> {
    let header = headers.get(axum::http::header::COOKIE)?;
    let value = header.to_str().ok()?;
    for pair in value.split(';') {
        let trimmed = pair.trim();
        let mut parts = trimmed.splitn(2, '=');
        let key = parts.next()?.trim();
        let val = parts.next()?.trim();
        if key == SESSION_COOKIE_NAME {
            return Some(val.to_string());
        }
    }
    None
}

async fn lookup_any_session(
    pool: &PgPool,
    token: &str,
) -> Result<Option<SessionRecord>, anyhow::Error> {
    let token_hash = hash_session_token(token);
    match SessionKind::from_token(token) {
        SessionKind::Full => lookup_full_session(pool, &token_hash).await,
        SessionKind::MfaBootstrap => lookup_mfa_bootstrap_session(pool, &token_hash).await,
        SessionKind::MfaChallenge => lookup_mfa_challenge_session(pool, &token_hash).await,
    }
}

async fn delete_any_session(pool: &PgPool, token: &str) -> Result<(), anyhow::Error> {
    let token_hash = hash_session_token(token);
    match SessionKind::from_token(token) {
        SessionKind::Full => delete_full_session(pool, &token_hash).await,
        SessionKind::MfaBootstrap => delete_mfa_bootstrap_session(pool, &token_hash).await,
        SessionKind::MfaChallenge => delete_mfa_challenge_session(pool, &token_hash).await,
    }
}